requirements.txt
streamlit
python-dotenv
requests
langchain
langchain-nvidia-ai-endpoints
pandas